feat: apply workspace group edit api
