+++ apps/web/src/schema/identity/workspace-group/api-verbs/update.ts
@@ -1,7 +1,4 @@
-import type { Tags } from '@/schema/_common/model';
-
 export interface WorkspaceGroupUpdateParameters {
     workspace_group_id: string;
     name?: string;
-    tags?: Tags;
 }
