+++ apps/web/src/services/advanced/components/WorkspaceGroupEditModal.vue
@@ -1,21 +1,48 @@
 <script setup lang="ts">
-import { reactive } from 'vue';
+import { reactive, watch } from 'vue';
 
+import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 import { PButtonModal, PFieldGroup, PTextInput } from '@cloudforet/mirinae';
+
+import type { WorkspaceGroupUpdateParameters } from '@/schema/identity/workspace-group/api-verbs/update';
+import type { WorkspaceGroupModel } from '@/schema/identity/workspace-group/model';
+
+import ErrorHandler from '@/common/composables/error/errorHandler';
 
 import { WORKSPACE_GROUP_MODAL_TYPE } from '@/services/advanced/constants/workspace-group-constant';
 import { useWorkspaceGroupPageStore } from '@/services/advanced/store/workspace-group-page-store';
 
+
 const workspaceGroupPageStore = useWorkspaceGroupPageStore();
 const workspaceGroupPageState = workspaceGroupPageStore.state;
+const workspaceGroupPageGetters = workspaceGroupPageStore.getters;
+
+const emit = defineEmits<{(e: 'confirm'): void; }>();
 
 const state = reactive({
     groupName: '',
     loading: false,
 });
 
-const handleConfirm = () => {
+const updateWorkspaceGroups = async () => {
+    state.loading = true;
+
+    try {
+        await SpaceConnector.clientV2.identity.workspaceGroup.update<WorkspaceGroupUpdateParameters, WorkspaceGroupModel>({
+            workspace_group_id: workspaceGroupPageGetters.selectedGroup.workspace_group_id,
+            name: state.groupName,
+        });
+    } catch (e) {
+        ErrorHandler.handleError(e);
+    } finally {
+        state.loading = false;
+    }
+};
+
+const handleConfirm = async () => {
+    await updateWorkspaceGroups();
     workspaceGroupPageStore.closeModal();
+    emit('confirm');
 };
 
 const handleCancel = () => {
@@ -26,6 +53,12 @@ const handleClose = () => {
     workspaceGroupPageStore.closeModal();
 };
 
+
+watch(() => workspaceGroupPageState.modal.visible, () => {
+    if (workspaceGroupPageState.modal.visible === WORKSPACE_GROUP_MODAL_TYPE.EDIT) {
+        state.groupName = workspaceGroupPageGetters.selectedGroup.name;
+    }
+});
 </script>
 
 <template>
@@ -45,7 +78,7 @@ const handleClose = () => {
                     :label="$t('IAM.WORKSPACE_GROUP.MODAL.EDIT_GROUP_NAME')"
                     style-type="secondary"
                 >
-                    <p-text-input :value="state.groupName"
+                    <p-text-input v-model="state.groupName"
                                   :placeholder="$t('IAM.WORKSPACE_GROUP.MODAL.EDIT_GROUP_NAME_PLACEHOLDER')"
                                   block
                     />
