+++ apps/web/src/services/advanced/pages/admin/AdminWorkspaceGroupPage.vue
@@ -1,7 +1,14 @@
 <script setup lang="ts">
-import { onUnmounted } from 'vue';
+import { onUnmounted, reactive } from 'vue';
 
+import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 import { PHorizontalLayout } from '@cloudforet/mirinae';
+
+import type { ListResponse } from '@/schema/_common/api-verbs/list';
+import type { WorkspaceGroupListParameters } from '@/schema/identity/workspace-group/api-verbs/list';
+import type { WorkspaceGroupModel } from '@/schema/identity/workspace-group/model';
+
+import ErrorHandler from '@/common/composables/error/errorHandler';
 
 import WorkspaceGroupAddUsersModal from '@/services/advanced/components/WorkspaceGroupAddUsersModal.vue';
 import WorkspaceGroupAddWorkspacesModal from '@/services/advanced/components/WorkspaceGroupAddWorkspacesModal.vue';
@@ -14,7 +21,43 @@ import WorkspaceGroupTab from '@/services/advanced/components/WorkspaceGroupTab.
 import WorkspaceGroupTable from '@/services/advanced/components/WorkspaceGroupTable.vue';
 import { useWorkspaceGroupPageStore } from '@/services/advanced/store/workspace-group-page-store';
 
+
+
 const workspaceGroupPageStore = useWorkspaceGroupPageStore();
+const workspaceGroupPageState = workspaceGroupPageStore.state;
+
+const state = reactive({
+    loading: false,
+});
+
+const fetchWorkspaceGroups = async () => {
+    state.loading = true;
+
+    try {
+        // TODO: apply Destructuring
+        const results = await SpaceConnector.clientV2.identity.workspaceGroup.list<WorkspaceGroupListParameters, ListResponse<WorkspaceGroupModel>>({}) as WorkspaceGroupModel[];
+
+        workspaceGroupPageState.groups = results;
+        workspaceGroupPageState.selectedIndices = [];
+    } catch (e) {
+        ErrorHandler.handleError(e);
+        workspaceGroupPageState.groups = [];
+    } finally {
+        state.loading = false;
+    }
+};
+
+const initWorkspaceGroups = async () => {
+    await fetchWorkspaceGroups();
+
+    workspaceGroupPageStore.$patch((_state) => {
+        _state.state.selectedIndices = [0];
+    });
+};
+
+(() => {
+    initWorkspaceGroups();
+})();
 
 onUnmounted(() => {
     workspaceGroupPageStore.$dispose();
@@ -32,7 +75,7 @@ onUnmounted(() => {
         </p-horizontal-layout>
         <workspace-group-tab />
         <workspace-group-create-modal />
-        <workspace-group-edit-modal />
+        <workspace-group-edit-modal @confirm="fetchWorkspaceGroups" />
         <workspace-group-delete-modal />
         <workspace-group-delete-status-modal />
         <workspace-group-add-workspaces-modal />
